[feat] signup 부분 리팩토링 완료
