+++ src/components/signup/types.tsx
@@ -1,0 +1,9 @@
+interface TermsProps {
+  isChecked: {
+    isCheckedTotal: boolean
+    isCheckedFirst: boolean
+    isCheckedSecond: boolean
+    isCheckedThird: boolean
+  }
+  setIsChecked: any
+}
